+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Fixed
 * Locales can now update `:code` value
 
+### Added
+* Added `#destroy` method to Locales
+
 ## 0.8.0
 ### Added
 * Added `:dynamic_entries` parameter on Client initialization